requirements.txt
fastapi
uvicorn[standard]
langgraph
langgraph-checkpoint-postgres
psycopg[binary,pool]
python-dotenv
langchain-anthropic
langchain-google-genai
pydantic
httpx